[REFACTOR] Reorganize directory structure
Flatten out an unnecessarily deep directory structure
