+++ command-expansion-server/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "build": "rm -rf ./build/packages/lib/CommandEDSLPreface.ts; tsc; cp ./src/packages/lib/CommandEDSLPreface.ts ./build/packages/lib/CommandEDSLPreface.ts;",
+    "build": "rm -rf ./build/lib/CommandEDSLPreface.ts; tsc; cp ./src/lib/CommandEDSLPreface.ts ./build/lib/CommandEDSLPreface.ts;",
     "start": "node ./build/app.js",
     "ts-watch": "tsc --watch",
     "dev-server": "supervisor -- --inspect=0.0.0.0 ./build/app.js",
