+++ command-expansion-server/src/lib/CodegenHelpers.ts
