+++ command-expansion-server/src/lib/getActivityTypescript.ts
@@ -1,6 +1,6 @@
 import { GraphQLClient, gql } from "graphql-request";
-import { globalDeclaration, indent, interfaceDeclaration } from "./packages/lib/CodegenHelpers.js";
-import { ErrorWithStatusCode } from "./utils/ErrorWithStatusCode.js";
+import { globalDeclaration, indent, interfaceDeclaration } from "./CodegenHelpers.js";
+import { ErrorWithStatusCode } from "../utils/ErrorWithStatusCode.js";
 import getLogger from "./utils/logger.js";
 
 const logger = getLogger("getActivityTypescript");
